+++ design-toolkit/components/src/components/accordion/index.tsx
@@ -71,7 +71,7 @@ export function Accordion({
   variant = 'cozy',
   ...props
 }: AccordionProps) {
-  // @ts-expect-error package version mismatch
+  // @ts-expect-error package version mismatch TODO
   const stateContext = useContext(AriaDisclosureGroupStateContext);
 
   return (
@@ -128,7 +128,7 @@ function AccordionHeader({ children }: AccordionHeaderProps) {
         ])}
         data-variant={variant}
       >
-        {/* @ts-expect-error package version mismatch */}
+        {/* @ts-expect-error package version mismatch TODO*/}
         <span
           className={IconButton.as({
             isDisabled,