+++ packages/predicates/src/is-greater/index.ts
@@ -11,20 +11,21 @@
  */
 
 /**
- * Determines if the first value is greater than the second value.
+ * Determines if one value is greater than another value.
+ * Right-to-left (composition) order.
  *
- * @param a - The first value to check against.
- * @param b - The second value to check against.
+ * @param a - The right hand value.
+ * @param b - The left hand value.
  * @template T - The type of the values.
  *
  * @remarks
  * pure function
  *
  * @example
- * isGreater(42)(23); // true
- * isGreater(5)(23); // false
+ * isGreater(42)(23); // false
+ * isGreater(5)(23); // true
  */
 export const isGreater =
   <T extends number | string>(a: T) =>
   (b: T) =>
-    a > b;
+    b > a;