+++ packages/predicates/src/is-lesser/index.ts
@@ -11,20 +11,21 @@
  */
 
 /**
- * Determines if the first value is lesser than the second value.
+ * Determines if one value is lesser than another value.
+ * Right-to-left (composition) order.
  *
- * @param a - The first value to check against.
- * @param b - The second value to check against.
+ * @param a - The right hand value.
+ * @param b - The left hand value.
  * @template T - The type of the values.
  *
  * @remarks
  * pure function
  *
  * @example
- * isLesser(42)(23); // false
- * isLesser(5)(23); // true
+ * isLesser(42)(23); // true
+ * isLesser(5)(23); // false
  */
 export const isLesser =
   <T>(a: T) =>
   (b: T) =>
-    a < b;
+    b < a;